@@ -2,54 +2,64 @@
   "name": "mitmproxy",
   "private": true,
   "scripts": {
-    "test": "jest ./src/js"
+    "test": "jest",
+    "build": "gulp prod",
+    "start": "gulp"
   },
   "jest": {
+    "testRegex": "__tests__/.*Spec.js$",
     "testPathDirs": [
-      "./src/js"
+      "<rootDir>/src/js"
     ],
-    "testPathIgnorePatterns": [
-      "/node_modules/",
-      "testutils.js"
-    ],
-    "testDirectoryName": "tests"
+    "unmockedModulePathPatterns": [
+      "react"
+    ]
   },
   "dependencies": {
-    "bootstrap": "^3.3.4",
+    "bootstrap": "^3.3.6",
+    "classnames": "^2.2.5",
     "flux": "^2.1.1",
-    "jquery": "^2.1.4",
-    "lodash": "^3.10.1",
-    "react": "^0.13.3",
-    "react-router": "^0.13.2"
+    "history": "^3.0.0",
+    "lodash": "^4.11.2",
+    "react": "^15.1.0",
+    "react-codemirror": "^0.2.6",
+    "react-dom": "^15.1.0",
+    "react-redux": "^4.4.5",
+    "redux": "^3.5.2",
+    "redux-logger": "^2.6.1",
+    "redux-thunk": "^2.1.0",
+    "shallowequal": "^0.2.2"
   },
   "devDependencies": {
-    "babelify": "^6.3.0",
-    "browserify": "^11.2.0",
-    "eslint": "^1.5.1",
-    "gulp": "^3.9.0",
-    "gulp-concat": "^2.6.0",
-    "gulp-connect": "^2.2.0",
-    "gulp-eslint": "^1.0.0",
-    "gulp-jshint": "^1.11.2",
-    "gulp-less": "^3.0.3",
+    "babel-core": "^6.7.7",
+    "babel-eslint": "^6.0.4",
+    "babel-jest": "^12.1.0",
+    "babel-plugin-transform-class-properties": "^6.6.0",
+    "babel-plugin-transform-object-rest-spread": "^6.8.0",
+    "babel-preset-es2015": "^6.6.0",
+    "babel-preset-react": "^6.5.0",
+    "babelify": "^7.3.0",
+    "browserify": "^13.0.0",
+    "envify": "^3.4.1",
+    "eslint": "^2.9.0",
+    "gulp": "^3.9.1",
+    "gulp-clean-css": "^2.0.6",
+    "gulp-eslint": "^2.0.0",
+    "gulp-less": "^3.0.5",
     "gulp-livereload": "^3.8.1",
-    "gulp-minify-css": "^1.2.1",
     "gulp-notify": "^2.2.0",
-    "gulp-peg": "^0.1.2",
-    "gulp-plumber": "^1.0.1",
-    "gulp-react": "^3.0.1",
+    "gulp-peg": "^0.2.0",
+    "gulp-plumber": "^1.1.0",
     "gulp-rename": "^1.2.2",
-    "gulp-replace": "^0.5.4",
     "gulp-sourcemaps": "^1.6.0",
-    "gulp-uglify": "^1.4.1",
-    "gulp-util": "^3.0.6",
-    "lodash": "^3.10.1",
-    "map-stream": "0.0.6",
-    "reactify": "^1.1.1",
+    "gulp-util": "^3.0.7",
+    "jest": "^12.1.1",
+    "react": "^15.2.1",
+    "react-addons-test-utils": "^15.2.1",
+    "react-dom": "^15.2.1",
     "uglifyify": "^3.0.1",
     "vinyl-buffer": "^1.0.0",
     "vinyl-source-stream": "^1.1.0",
-    "vinyl-transform": "^1.0.0",
-    "watchify": "^3.4.0"
+    "watchify": "^3.7.0"
   }
 }
