@@ -21,7 +21,7 @@ import shlex
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 sys.path.insert(0, os.path.abspath('..'))
 
-import libmproxy.version
+import netlib.version
 
 # -- General configuration ------------------------------------------------
 
@@ -46,7 +46,7 @@ napoleon_include_special_with_doc = False
 autodoc_member_order = "bysource"
 
 # Add any paths that contain templates here, relative to this directory.
-#templates_path = ['_templates']
+templates_path = ['_templates']
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
@@ -69,9 +69,9 @@ author = u'The mitmproxy project'
 # built documents.
 #
 # The short X.Y version.
-version = libmproxy.version.VERSION
+version = netlib.version.VERSION
 # The full version, including alpha/beta/rc tags.
-release = libmproxy.version.VERSION
+release = netlib.version.VERSION
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
@@ -109,7 +109,7 @@ exclude_patterns = ['_build']
 pygments_style = 'sphinx'
 
 # A list of ignored prefixes for module index sorting.
-modindex_common_prefix = ['libmproxy.']
+modindex_common_prefix = ['mitmproxy.']
 
 # If true, keep warnings as "system message" paragraphs in the built documents.
 #keep_warnings = False
@@ -128,7 +128,7 @@ html_theme = 'sphinx_rtd_theme'
 # further.  For a list of options available for each theme, see the
 # documentation.
 html_theme_options = {
-    # 'logo_only': True,
+    'logo_only': True,
 }
 
 # Add any paths that contain custom themes here, relative to this directory.
@@ -143,7 +143,7 @@ html_title = "mitmproxy %s documentation" % version
 
 # The name of an image file (relative to this directory) to place at the top
 # of the sidebar.
-html_logo = "mitmproxy-long.png"
+html_logo = "mitmproxy-docs.png"
 
 # The name of an image file (within the static path) to use as favicon of the
 # docs.  This file should be a Windows icon file (.ico) being 16x16 or 32x32
@@ -153,7 +153,7 @@ html_favicon = "favicon.ico"
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-# html_static_path = ['_static']
+html_static_path = ['_static']
 
 # Add any extra paths that contain custom files (such as robots.txt or
 # .htaccess) here, relative to this directory. These files are copied
@@ -217,3 +217,6 @@ html_favicon = "favicon.ico"
 
 # Output file base name for HTML help builder.
 htmlhelp_basename = 'mitmproxydoc'
+
+def setup(app):
+    app.add_stylesheet('theme_overrides.css')
