@@ -1,4 +1,7 @@
-def request(context, flow):
-    f = context.duplicate_flow(flow)
+from mitmproxy import master
+
+
+def request(flow):
+    f = master.duplicate_flow(flow)
     f.request.path = "/changed"
-    context.replay_request(f)
+    master.replay_request(f, block=True, run_scripthooks=False)
