@@ -1,16 +1,20 @@
-# This scripts demonstrates how to use mitmproxy's filter pattern in inline scripts.
+# This scripts demonstrates how to use mitmproxy's filter pattern in scripts.
 # Usage: mitmdump -s "filt.py FILTER"
+import sys
+from mitmproxy import filt
 
-from libmproxy import filt
+
+class Filter:
+    def __init__(self, spec):
+        self.filter = filt.parse(spec)
+
+    def response(self, flow):
+        if flow.match(self.filter):
+            print("Flow matches filter:")
+            print(flow)
 
 
-def start(context, argv):
-    if len(argv) != 2:
+def start():
+    if len(sys.argv) != 2:
         raise ValueError("Usage: -s 'filt.py FILTER'")
-    context.filter = filt.parse(argv[1])
-
-
-def response(context, flow):
-    if flow.match(context.filter):
-        print("Flow matches filter:")
-        print(flow)
+    return Filter(sys.argv[1])
