@@ -1,20 +1,23 @@
 import random
 import sys
 
-from libmproxy.flow import FlowWriter
+from mitmproxy.flow import FlowWriter
 
 
-def start(context, argv):
-    if len(argv) != 2:
+class Writer:
+    def __init__(self, path):
+        if path == "-":
+            f = sys.stdout
+        else:
+            f = open(path, "wb")
+        self.w = FlowWriter(f)
+
+    def response(self, flow):
+        if random.choice([True, False]):
+            self.w.add(flow)
+
+
+def start():
+    if len(sys.argv) != 2:
         raise ValueError('Usage: -s "flowriter.py filename"')
-
-    if argv[1] == "-":
-        f = sys.stdout
-    else:
-        f = open(argv[1], "wb")
-    context.flow_writer = FlowWriter(f)
-
-
-def response(context, flow):
-    if random.choice([True, False]):
-        context.flow_writer.add(flow)
+    return Writer(sys.argv[1])
