@@ -16,7 +16,6 @@ import sys
 
 
 class Wrapper(object):
-
     def __init__(self, port, extra_arguments=None):
         self.port = port
         self.extra_arguments = extra_arguments
@@ -142,7 +141,7 @@ class Wrapper(object):
             '--toggle',
             action='store_true',
             help='just toggle the proxy configuration')
-#         parser.add_argument('--honeyproxy', action='store_true', help='run honeyproxy instead of mitmproxy')
+        # parser.add_argument('--honeyproxy', action='store_true', help='run honeyproxy instead of mitmproxy')
         parser.add_argument(
             '-p',
             '--port',
@@ -155,8 +154,8 @@ class Wrapper(object):
 
         if args.toggle:
             wrapper.toggle_proxy()
-#         elif args.honeyproxy:
-#             wrapper.wrap_honeyproxy()
+        # elif args.honeyproxy:
+        #     wrapper.wrap_honeyproxy()
         else:
             wrapper.wrap_mitmproxy()
 
