@@ -1,6 +1,2 @@
-
-def request(context, flow):
-    q = flow.request.get_query()
-    if q:
-        q["mitmproxy"] = ["rocks"]
-        flow.request.set_query(q)
+def request(flow):
+    flow.request.query["mitmproxy"] = "rocks"
