@@ -1,18 +1,21 @@
 # Usage: mitmdump -s "modify_response_body.py mitmproxy bananas"
 # (this script works best with --anticache)
-from libmproxy.models import decoded
+import sys
 
 
-def start(context, argv):
-    if len(argv) != 3:
-        raise ValueError('Usage: -s "modify-response-body.py old new"')
+state = {}
+
+
+def start():
+    if len(sys.argv) != 3:
+        raise ValueError('Usage: -s "modify_response_body.py old new"')
     # You may want to use Python's argparse for more sophisticated argument
     # parsing.
-    context.old, context.new = argv[1], argv[2]
+    state["old"], state["new"] = sys.argv[1].encode(), sys.argv[2].encode()
 
 
-def response(context, flow):
-    with decoded(flow.response):  # automatically decode gzipped responses.
-        flow.response.content = flow.response.content.replace(
-            context.old,
-            context.new)
+def response(flow):
+    flow.response.content = flow.response.content.replace(
+        state["old"],
+        state["new"]
+    )
