@@ -1,9 +1,10 @@
 import time
-from libmproxy.script import concurrent
+import mitmproxy
+from mitmproxy.script import concurrent
 
 
 @concurrent  # Remove this and see what happens
-def request(context, flow):
-    print("handle request: %s%s" % (flow.request.host, flow.request.path))
+def request(flow):
+    mitmproxy.ctx.log("handle request: %s%s" % (flow.request.host, flow.request.path))
     time.sleep(5)
-    print("start  request: %s%s" % (flow.request.host, flow.request.path))
+    mitmproxy.ctx.log("start  request: %s%s" % (flow.request.host, flow.request.path))
