@@ -5,29 +5,23 @@ var conf = require('./conf.js');
 
 // Sorted alphabetically!
 var babelify = require('babelify');
+var envify = require('envify/custom');
 var browserify = require('browserify');
 var gulp = require("gulp");
-var concat = require('gulp-concat');
-var connect = require('gulp-connect');
 var eslint = require('gulp-eslint');
 var less = require("gulp-less");
 var livereload = require("gulp-livereload");
-var minifyCSS = require('gulp-minify-css');
+var cleanCSS = require('gulp-clean-css');
 var notify = require("gulp-notify");
 var peg = require("gulp-peg");
 var plumber = require("gulp-plumber");
-var react = require("gulp-react");
 var rename = require("gulp-rename");
-var replace = require('gulp-replace');
 var sourcemaps = require('gulp-sourcemaps');
 var gutil = require("gulp-util");
 var _ = require('lodash');
-var map = require("map-stream");
-var reactify = require('reactify');
 var uglifyify = require('uglifyify');
 var buffer = require('vinyl-buffer');
 var source = require('vinyl-source-stream');
-var transform = require('vinyl-transform');
 var watchify = require('watchify');
 
 var vendor_packages = _.difference(
@@ -75,7 +69,7 @@ function styles(files, dev){
         .pipe(dev ? plumber(handleError) : gutil.noop())
         .pipe(sourcemaps.init())
         .pipe(less())
-        .pipe(dev ? gutil.noop() : minifyCSS())
+        .pipe(dev ? gutil.noop() : cleanCSS())
         .pipe(sourcemaps.write(".", {sourceRoot: fixLessSourceMaps}))
         .pipe(gulp.dest(conf.static))
         .pipe(livereload({auto: false}));
@@ -98,11 +92,16 @@ function buildScript(bundler, filename, dev) {
     if (dev) {
         bundler = watchify(bundler);
     } else {
+        bundler = bundler.transform(envify({ _: 'purge', NODE_ENV: 'production' }), { global: true });
         bundler = bundler.transform({global: true}, uglifyify);
     }
 
     function rebundle() {
         return bundler.bundle()
+            .on('error', function(error) {
+                gutil.log(error + '\n' + error.codeFrame);
+                this.emit('end');
+            })
             .pipe(dev ? plumber(handleError) : gutil.noop())
             .pipe(source('bundle.js'))
             .pipe(buffer())
@@ -116,6 +115,7 @@ function buildScript(bundler, filename, dev) {
     // listen for an update and run rebundle
     bundler.on('update', rebundle);
     bundler.on('log', gutil.log);
+    bundler.on('error', gutil.log);
 
     // run it once the first time buildScript is called
     return rebundle();
@@ -144,13 +144,14 @@ function app_stream(dev) {
     var bundler = browserify({
         entries: [conf.js.app],
         debug: true,
+        extensions: ['.jsx'],
         cache: {}, // required for watchify
         packageCache: {} // required for watchify
     });
     for (var vp of vendor_packages) {
         bundler.external(vp);
     }
-    bundler = bundler.transform(babelify).transform(reactify);
+    bundler = bundler.transform(babelify);
     return buildScript(bundler, "app.js", dev);
 }
 gulp.task('scripts-app-dev', function () {
