@@ -1,8 +1,10 @@
 module.exports = (function() {
+  "use strict";
+
   /*
-   * Generated by PEG.js 0.8.0.
+   * Generated by PEG.js 0.9.0.
    *
-   * http://pegjs.majda.cz/
+   * http://pegjs.org/
    */
 
   function peg$subclass(child, parent) {
@@ -11,21 +13,23 @@ module.exports = (function() {
     child.prototype = new ctor();
   }
 
-  function SyntaxError(message, expected, found, offset, line, column) {
+  function peg$SyntaxError(message, expected, found, location) {
     this.message  = message;
     this.expected = expected;
     this.found    = found;
-    this.offset   = offset;
-    this.line     = line;
-    this.column   = column;
-
+    this.location = location;
     this.name     = "SyntaxError";
+
+    if (typeof Error.captureStackTrace === "function") {
+      Error.captureStackTrace(this, peg$SyntaxError);
+    }
   }
 
-  peg$subclass(SyntaxError, Error);
+  peg$subclass(peg$SyntaxError, Error);
 
-  function parse(input) {
+  function peg$parse(input) {
     var options = arguments.length > 1 ? arguments[1] : {},
+        parser  = this,
 
         peg$FAILED = {},
 
@@ -33,117 +37,111 @@ module.exports = (function() {
         peg$startRuleFunction  = peg$parsestart,
 
         peg$c0 = { type: "other", description: "filter expression" },
-        peg$c1 = peg$FAILED,
-        peg$c2 = function(orExpr) { return orExpr; },
-        peg$c3 = [],
-        peg$c4 = function() {return trueFilter; },
-        peg$c5 = { type: "other", description: "whitespace" },
-        peg$c6 = /^[ \t\n\r]/,
-        peg$c7 = { type: "class", value: "[ \\t\\n\\r]", description: "[ \\t\\n\\r]" },
-        peg$c8 = { type: "other", description: "control character" },
-        peg$c9 = /^[|&!()~"]/,
-        peg$c10 = { type: "class", value: "[|&!()~\"]", description: "[|&!()~\"]" },
-        peg$c11 = { type: "other", description: "optional whitespace" },
-        peg$c12 = "|",
-        peg$c13 = { type: "literal", value: "|", description: "\"|\"" },
-        peg$c14 = function(first, second) { return or(first, second); },
-        peg$c15 = "&",
-        peg$c16 = { type: "literal", value: "&", description: "\"&\"" },
-        peg$c17 = function(first, second) { return and(first, second); },
-        peg$c18 = "!",
-        peg$c19 = { type: "literal", value: "!", description: "\"!\"" },
-        peg$c20 = function(expr) { return not(expr); },
-        peg$c21 = "(",
-        peg$c22 = { type: "literal", value: "(", description: "\"(\"" },
-        peg$c23 = ")",
-        peg$c24 = { type: "literal", value: ")", description: "\")\"" },
-        peg$c25 = function(expr) { return binding(expr); },
-        peg$c26 = "~a",
-        peg$c27 = { type: "literal", value: "~a", description: "\"~a\"" },
-        peg$c28 = function() { return assetFilter; },
-        peg$c29 = "~e",
-        peg$c30 = { type: "literal", value: "~e", description: "\"~e\"" },
-        peg$c31 = function() { return errorFilter; },
-        peg$c32 = "~q",
-        peg$c33 = { type: "literal", value: "~q", description: "\"~q\"" },
-        peg$c34 = function() { return noResponseFilter; },
-        peg$c35 = "~s",
-        peg$c36 = { type: "literal", value: "~s", description: "\"~s\"" },
-        peg$c37 = function() { return responseFilter; },
-        peg$c38 = "true",
-        peg$c39 = { type: "literal", value: "true", description: "\"true\"" },
-        peg$c40 = function() { return trueFilter; },
-        peg$c41 = "false",
-        peg$c42 = { type: "literal", value: "false", description: "\"false\"" },
-        peg$c43 = function() { return falseFilter; },
-        peg$c44 = "~c",
-        peg$c45 = { type: "literal", value: "~c", description: "\"~c\"" },
-        peg$c46 = function(s) { return responseCode(s); },
-        peg$c47 = "~d",
-        peg$c48 = { type: "literal", value: "~d", description: "\"~d\"" },
-        peg$c49 = function(s) { return domain(s); },
-        peg$c50 = "~h",
-        peg$c51 = { type: "literal", value: "~h", description: "\"~h\"" },
-        peg$c52 = function(s) { return header(s); },
-        peg$c53 = "~hq",
-        peg$c54 = { type: "literal", value: "~hq", description: "\"~hq\"" },
-        peg$c55 = function(s) { return requestHeader(s); },
-        peg$c56 = "~hs",
-        peg$c57 = { type: "literal", value: "~hs", description: "\"~hs\"" },
-        peg$c58 = function(s) { return responseHeader(s); },
-        peg$c59 = "~m",
-        peg$c60 = { type: "literal", value: "~m", description: "\"~m\"" },
-        peg$c61 = function(s) { return method(s); },
-        peg$c62 = "~t",
-        peg$c63 = { type: "literal", value: "~t", description: "\"~t\"" },
-        peg$c64 = function(s) { return contentType(s); },
-        peg$c65 = "~tq",
-        peg$c66 = { type: "literal", value: "~tq", description: "\"~tq\"" },
-        peg$c67 = function(s) { return requestContentType(s); },
-        peg$c68 = "~ts",
-        peg$c69 = { type: "literal", value: "~ts", description: "\"~ts\"" },
-        peg$c70 = function(s) { return responseContentType(s); },
-        peg$c71 = "~u",
-        peg$c72 = { type: "literal", value: "~u", description: "\"~u\"" },
-        peg$c73 = function(s) { return url(s); },
-        peg$c74 = { type: "other", description: "integer" },
-        peg$c75 = null,
-        peg$c76 = /^['"]/,
-        peg$c77 = { type: "class", value: "['\"]", description: "['\"]" },
-        peg$c78 = /^[0-9]/,
-        peg$c79 = { type: "class", value: "[0-9]", description: "[0-9]" },
-        peg$c80 = function(digits) { return parseInt(digits.join(""), 10); },
-        peg$c81 = { type: "other", description: "string" },
-        peg$c82 = "\"",
-        peg$c83 = { type: "literal", value: "\"", description: "\"\\\"\"" },
-        peg$c84 = function(chars) { return chars.join(""); },
-        peg$c85 = "'",
-        peg$c86 = { type: "literal", value: "'", description: "\"'\"" },
-        peg$c87 = void 0,
-        peg$c88 = /^["\\]/,
-        peg$c89 = { type: "class", value: "[\"\\\\]", description: "[\"\\\\]" },
-        peg$c90 = { type: "any", description: "any character" },
-        peg$c91 = function(char) { return char; },
-        peg$c92 = "\\",
-        peg$c93 = { type: "literal", value: "\\", description: "\"\\\\\"" },
-        peg$c94 = /^['\\]/,
-        peg$c95 = { type: "class", value: "['\\\\]", description: "['\\\\]" },
-        peg$c96 = /^['"\\]/,
-        peg$c97 = { type: "class", value: "['\"\\\\]", description: "['\"\\\\]" },
-        peg$c98 = "n",
-        peg$c99 = { type: "literal", value: "n", description: "\"n\"" },
-        peg$c100 = function() { return "\n"; },
-        peg$c101 = "r",
-        peg$c102 = { type: "literal", value: "r", description: "\"r\"" },
-        peg$c103 = function() { return "\r"; },
-        peg$c104 = "t",
-        peg$c105 = { type: "literal", value: "t", description: "\"t\"" },
-        peg$c106 = function() { return "\t"; },
+        peg$c1 = function(orExpr) { return orExpr; },
+        peg$c2 = { type: "other", description: "whitespace" },
+        peg$c3 = /^[ \t\n\r]/,
+        peg$c4 = { type: "class", value: "[ \\t\\n\\r]", description: "[ \\t\\n\\r]" },
+        peg$c5 = { type: "other", description: "control character" },
+        peg$c6 = /^[|&!()~"]/,
+        peg$c7 = { type: "class", value: "[|&!()~\"]", description: "[|&!()~\"]" },
+        peg$c8 = { type: "other", description: "optional whitespace" },
+        peg$c9 = "|",
+        peg$c10 = { type: "literal", value: "|", description: "\"|\"" },
+        peg$c11 = function(first, second) { return or(first, second); },
+        peg$c12 = "&",
+        peg$c13 = { type: "literal", value: "&", description: "\"&\"" },
+        peg$c14 = function(first, second) { return and(first, second); },
+        peg$c15 = "!",
+        peg$c16 = { type: "literal", value: "!", description: "\"!\"" },
+        peg$c17 = function(expr) { return not(expr); },
+        peg$c18 = "(",
+        peg$c19 = { type: "literal", value: "(", description: "\"(\"" },
+        peg$c20 = ")",
+        peg$c21 = { type: "literal", value: ")", description: "\")\"" },
+        peg$c22 = function(expr) { return binding(expr); },
+        peg$c23 = "~a",
+        peg$c24 = { type: "literal", value: "~a", description: "\"~a\"" },
+        peg$c25 = function() { return assetFilter; },
+        peg$c26 = "~e",
+        peg$c27 = { type: "literal", value: "~e", description: "\"~e\"" },
+        peg$c28 = function() { return errorFilter; },
+        peg$c29 = "~q",
+        peg$c30 = { type: "literal", value: "~q", description: "\"~q\"" },
+        peg$c31 = function() { return noResponseFilter; },
+        peg$c32 = "~s",
+        peg$c33 = { type: "literal", value: "~s", description: "\"~s\"" },
+        peg$c34 = function() { return responseFilter; },
+        peg$c35 = "true",
+        peg$c36 = { type: "literal", value: "true", description: "\"true\"" },
+        peg$c37 = function() { return trueFilter; },
+        peg$c38 = "false",
+        peg$c39 = { type: "literal", value: "false", description: "\"false\"" },
+        peg$c40 = function() { return falseFilter; },
+        peg$c41 = "~c",
+        peg$c42 = { type: "literal", value: "~c", description: "\"~c\"" },
+        peg$c43 = function(s) { return responseCode(s); },
+        peg$c44 = "~d",
+        peg$c45 = { type: "literal", value: "~d", description: "\"~d\"" },
+        peg$c46 = function(s) { return domain(s); },
+        peg$c47 = "~h",
+        peg$c48 = { type: "literal", value: "~h", description: "\"~h\"" },
+        peg$c49 = function(s) { return header(s); },
+        peg$c50 = "~hq",
+        peg$c51 = { type: "literal", value: "~hq", description: "\"~hq\"" },
+        peg$c52 = function(s) { return requestHeader(s); },
+        peg$c53 = "~hs",
+        peg$c54 = { type: "literal", value: "~hs", description: "\"~hs\"" },
+        peg$c55 = function(s) { return responseHeader(s); },
+        peg$c56 = "~m",
+        peg$c57 = { type: "literal", value: "~m", description: "\"~m\"" },
+        peg$c58 = function(s) { return method(s); },
+        peg$c59 = "~t",
+        peg$c60 = { type: "literal", value: "~t", description: "\"~t\"" },
+        peg$c61 = function(s) { return contentType(s); },
+        peg$c62 = "~tq",
+        peg$c63 = { type: "literal", value: "~tq", description: "\"~tq\"" },
+        peg$c64 = function(s) { return requestContentType(s); },
+        peg$c65 = "~ts",
+        peg$c66 = { type: "literal", value: "~ts", description: "\"~ts\"" },
+        peg$c67 = function(s) { return responseContentType(s); },
+        peg$c68 = "~u",
+        peg$c69 = { type: "literal", value: "~u", description: "\"~u\"" },
+        peg$c70 = function(s) { return url(s); },
+        peg$c71 = { type: "other", description: "integer" },
+        peg$c72 = /^['"]/,
+        peg$c73 = { type: "class", value: "['\"]", description: "['\"]" },
+        peg$c74 = /^[0-9]/,
+        peg$c75 = { type: "class", value: "[0-9]", description: "[0-9]" },
+        peg$c76 = function(digits) { return parseInt(digits.join(""), 10); },
+        peg$c77 = { type: "other", description: "string" },
+        peg$c78 = "\"",
+        peg$c79 = { type: "literal", value: "\"", description: "\"\\\"\"" },
+        peg$c80 = function(chars) { return chars.join(""); },
+        peg$c81 = "'",
+        peg$c82 = { type: "literal", value: "'", description: "\"'\"" },
+        peg$c83 = /^["\\]/,
+        peg$c84 = { type: "class", value: "[\"\\\\]", description: "[\"\\\\]" },
+        peg$c85 = { type: "any", description: "any character" },
+        peg$c86 = function(char) { return char; },
+        peg$c87 = "\\",
+        peg$c88 = { type: "literal", value: "\\", description: "\"\\\\\"" },
+        peg$c89 = /^['\\]/,
+        peg$c90 = { type: "class", value: "['\\\\]", description: "['\\\\]" },
+        peg$c91 = /^['"\\]/,
+        peg$c92 = { type: "class", value: "['\"\\\\]", description: "['\"\\\\]" },
+        peg$c93 = "n",
+        peg$c94 = { type: "literal", value: "n", description: "\"n\"" },
+        peg$c95 = function() { return "\n"; },
+        peg$c96 = "r",
+        peg$c97 = { type: "literal", value: "r", description: "\"r\"" },
+        peg$c98 = function() { return "\r"; },
+        peg$c99 = "t",
+        peg$c100 = { type: "literal", value: "t", description: "\"t\"" },
+        peg$c101 = function() { return "\t"; },
 
         peg$currPos          = 0,
-        peg$reportedPos      = 0,
-        peg$cachedPos        = 0,
-        peg$cachedPosDetails = { line: 1, column: 1, seenCR: false },
+        peg$savedPos         = 0,
+        peg$posDetailsCache  = [{ line: 1, column: 1, seenCR: false }],
         peg$maxFailPos       = 0,
         peg$maxFailExpected  = [],
         peg$silentFails      = 0,
@@ -159,38 +157,51 @@ module.exports = (function() {
     }
 
     function text() {
-      return input.substring(peg$reportedPos, peg$currPos);
+      return input.substring(peg$savedPos, peg$currPos);
     }
 
-    function offset() {
-      return peg$reportedPos;
-    }
-
-    function line() {
-      return peg$computePosDetails(peg$reportedPos).line;
-    }
-
-    function column() {
-      return peg$computePosDetails(peg$reportedPos).column;
+    function location() {
+      return peg$computeLocation(peg$savedPos, peg$currPos);
     }
 
     function expected(description) {
       throw peg$buildException(
         null,
         [{ type: "other", description: description }],
-        peg$reportedPos
+        input.substring(peg$savedPos, peg$currPos),
+        peg$computeLocation(peg$savedPos, peg$currPos)
       );
     }
 
     function error(message) {
-      throw peg$buildException(message, null, peg$reportedPos);
+      throw peg$buildException(
+        message,
+        null,
+        input.substring(peg$savedPos, peg$currPos),
+        peg$computeLocation(peg$savedPos, peg$currPos)
+      );
     }
 
     function peg$computePosDetails(pos) {
-      function advance(details, startPos, endPos) {
-        var p, ch;
+      var details = peg$posDetailsCache[pos],
+          p, ch;
 
-        for (p = startPos; p < endPos; p++) {
+      if (details) {
+        return details;
+      } else {
+        p = pos - 1;
+        while (!peg$posDetailsCache[p]) {
+          p--;
+        }
+
+        details = peg$posDetailsCache[p];
+        details = {
+          line:   details.line,
+          column: details.column,
+          seenCR: details.seenCR
+        };
+
+        while (p < pos) {
           ch = input.charAt(p);
           if (ch === "\n") {
             if (!details.seenCR) { details.line++; }
@@ -204,19 +215,31 @@ module.exports = (function() {
             details.column++;
             details.seenCR = false;
           }
-        }
-      }
 
-      if (peg$cachedPos !== pos) {
-        if (peg$cachedPos > pos) {
-          peg$cachedPos = 0;
-          peg$cachedPosDetails = { line: 1, column: 1, seenCR: false };
+          p++;
         }
-        advance(peg$cachedPosDetails, peg$cachedPos, pos);
-        peg$cachedPos = pos;
-      }
 
-      return peg$cachedPosDetails;
+        peg$posDetailsCache[pos] = details;
+        return details;
+      }
+    }
+
+    function peg$computeLocation(startPos, endPos) {
+      var startPosDetails = peg$computePosDetails(startPos),
+          endPosDetails   = peg$computePosDetails(endPos);
+
+      return {
+        start: {
+          offset: startPos,
+          line:   startPosDetails.line,
+          column: startPosDetails.column
+        },
+        end: {
+          offset: endPos,
+          line:   endPosDetails.line,
+          column: endPosDetails.column
+        }
+      };
     }
 
     function peg$fail(expected) {
@@ -230,7 +253,7 @@ module.exports = (function() {
       peg$maxFailExpected.push(expected);
     }
 
-    function peg$buildException(message, expected, pos) {
+    function peg$buildException(message, expected, found, location) {
       function cleanupExpected(expected) {
         var i = 1;
 
@@ -267,8 +290,8 @@ module.exports = (function() {
             .replace(/\r/g,   '\\r')
             .replace(/[\x00-\x07\x0B\x0E\x0F]/g, function(ch) { return '\\x0' + hex(ch); })
             .replace(/[\x10-\x1F\x80-\xFF]/g,    function(ch) { return '\\x'  + hex(ch); })
-            .replace(/[\u0180-\u0FFF]/g,         function(ch) { return '\\u0' + hex(ch); })
-            .replace(/[\u1080-\uFFFF]/g,         function(ch) { return '\\u'  + hex(ch); });
+            .replace(/[\u0100-\u0FFF]/g,         function(ch) { return '\\u0' + hex(ch); })
+            .replace(/[\u1000-\uFFFF]/g,         function(ch) { return '\\u'  + hex(ch); });
         }
 
         var expectedDescs = new Array(expected.length),
@@ -289,20 +312,15 @@ module.exports = (function() {
         return "Expected " + expectedDesc + " but " + foundDesc + " found.";
       }
 
-      var posDetails = peg$computePosDetails(pos),
-          found      = pos < input.length ? input.charAt(pos) : null;
-
       if (expected !== null) {
         cleanupExpected(expected);
       }
 
-      return new SyntaxError(
+      return new peg$SyntaxError(
         message !== null ? message : buildMessage(expected, found),
         expected,
         found,
-        pos,
-        posDetails.line,
-        posDetails.column
+        location
       );
     }
 
@@ -317,29 +335,20 @@ module.exports = (function() {
         if (s2 !== peg$FAILED) {
           s3 = peg$parse__();
           if (s3 !== peg$FAILED) {
-            peg$reportedPos = s0;
-            s1 = peg$c2(s2);
+            peg$savedPos = s0;
+            s1 = peg$c1(s2);
             s0 = s1;
           } else {
             peg$currPos = s0;
-            s0 = peg$c1;
+            s0 = peg$FAILED;
           }
         } else {
           peg$currPos = s0;
-          s0 = peg$c1;
+          s0 = peg$FAILED;
         }
       } else {
         peg$currPos = s0;
-        s0 = peg$c1;
-      }
-      if (s0 === peg$FAILED) {
-        s0 = peg$currPos;
-        s1 = [];
-        if (s1 !== peg$FAILED) {
-          peg$reportedPos = s0;
-          s1 = peg$c4();
-        }
-        s0 = s1;
+        s0 = peg$FAILED;
       }
       peg$silentFails--;
       if (s0 === peg$FAILED) {
@@ -351,6 +360,26 @@ module.exports = (function() {
     }
 
     function peg$parsews() {
+      var s0, s1;
+
+      peg$silentFails++;
+      if (peg$c3.test(input.charAt(peg$currPos))) {
+        s0 = input.charAt(peg$currPos);
+        peg$currPos++;
+      } else {
+        s0 = peg$FAILED;
+        if (peg$silentFails === 0) { peg$fail(peg$c4); }
+      }
+      peg$silentFails--;
+      if (s0 === peg$FAILED) {
+        s1 = peg$FAILED;
+        if (peg$silentFails === 0) { peg$fail(peg$c2); }
+      }
+
+      return s0;
+    }
+
+    function peg$parsecc() {
       var s0, s1;
 
       peg$silentFails++;
@@ -370,26 +399,6 @@ module.exports = (function() {
       return s0;
     }
 
-    function peg$parsecc() {
-      var s0, s1;
-
-      peg$silentFails++;
-      if (peg$c9.test(input.charAt(peg$currPos))) {
-        s0 = input.charAt(peg$currPos);
-        peg$currPos++;
-      } else {
-        s0 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c10); }
-      }
-      peg$silentFails--;
-      if (s0 === peg$FAILED) {
-        s1 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c8); }
-      }
-
-      return s0;
-    }
-
     function peg$parse__() {
       var s0, s1;
 
@@ -403,7 +412,7 @@ module.exports = (function() {
       peg$silentFails--;
       if (s0 === peg$FAILED) {
         s1 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c11); }
+        if (peg$silentFails === 0) { peg$fail(peg$c8); }
       }
 
       return s0;
@@ -418,39 +427,39 @@ module.exports = (function() {
         s2 = peg$parse__();
         if (s2 !== peg$FAILED) {
           if (input.charCodeAt(peg$currPos) === 124) {
-            s3 = peg$c12;
+            s3 = peg$c9;
             peg$currPos++;
           } else {
             s3 = peg$FAILED;
-            if (peg$silentFails === 0) { peg$fail(peg$c13); }
+            if (peg$silentFails === 0) { peg$fail(peg$c10); }
           }
           if (s3 !== peg$FAILED) {
             s4 = peg$parse__();
             if (s4 !== peg$FAILED) {
               s5 = peg$parseOrExpr();
               if (s5 !== peg$FAILED) {
-                peg$reportedPos = s0;
-                s1 = peg$c14(s1, s5);
+                peg$savedPos = s0;
+                s1 = peg$c11(s1, s5);
                 s0 = s1;
               } else {
                 peg$currPos = s0;
-                s0 = peg$c1;
+                s0 = peg$FAILED;
               }
             } else {
               peg$currPos = s0;
-              s0 = peg$c1;
+              s0 = peg$FAILED;
             }
           } else {
             peg$currPos = s0;
-            s0 = peg$c1;
+            s0 = peg$FAILED;
           }
         } else {
           peg$currPos = s0;
-          s0 = peg$c1;
+          s0 = peg$FAILED;
         }
       } else {
         peg$currPos = s0;
-        s0 = peg$c1;
+        s0 = peg$FAILED;
       }
       if (s0 === peg$FAILED) {
         s0 = peg$parseAndExpr();
@@ -468,39 +477,39 @@ module.exports = (function() {
         s2 = peg$parse__();
         if (s2 !== peg$FAILED) {
           if (input.charCodeAt(peg$currPos) === 38) {
-            s3 = peg$c15;
+            s3 = peg$c12;
             peg$currPos++;
           } else {
             s3 = peg$FAILED;
-            if (peg$silentFails === 0) { peg$fail(peg$c16); }
+            if (peg$silentFails === 0) { peg$fail(peg$c13); }
           }
           if (s3 !== peg$FAILED) {
             s4 = peg$parse__();
             if (s4 !== peg$FAILED) {
               s5 = peg$parseAndExpr();
               if (s5 !== peg$FAILED) {
-                peg$reportedPos = s0;
-                s1 = peg$c17(s1, s5);
+                peg$savedPos = s0;
+                s1 = peg$c14(s1, s5);
                 s0 = s1;
               } else {
                 peg$currPos = s0;
-                s0 = peg$c1;
+                s0 = peg$FAILED;
               }
             } else {
               peg$currPos = s0;
-              s0 = peg$c1;
+              s0 = peg$FAILED;
             }
           } else {
             peg$currPos = s0;
-            s0 = peg$c1;
+            s0 = peg$FAILED;
           }
         } else {
           peg$currPos = s0;
-          s0 = peg$c1;
+          s0 = peg$FAILED;
         }
       } else {
         peg$currPos = s0;
-        s0 = peg$c1;
+        s0 = peg$FAILED;
       }
       if (s0 === peg$FAILED) {
         s0 = peg$currPos;
@@ -514,25 +523,25 @@ module.exports = (function() {
               s3 = peg$parsews();
             }
           } else {
-            s2 = peg$c1;
+            s2 = peg$FAILED;
           }
           if (s2 !== peg$FAILED) {
             s3 = peg$parseAndExpr();
             if (s3 !== peg$FAILED) {
-              peg$reportedPos = s0;
-              s1 = peg$c17(s1, s3);
+              peg$savedPos = s0;
+              s1 = peg$c14(s1, s3);
               s0 = s1;
             } else {
               peg$currPos = s0;
-              s0 = peg$c1;
+              s0 = peg$FAILED;
             }
           } else {
             peg$currPos = s0;
-            s0 = peg$c1;
+            s0 = peg$FAILED;
           }
         } else {
           peg$currPos = s0;
-          s0 = peg$c1;
+          s0 = peg$FAILED;
         }
         if (s0 === peg$FAILED) {
           s0 = peg$parseNotExpr();
@@ -547,31 +556,31 @@ module.exports = (function() {
 
       s0 = peg$currPos;
       if (input.charCodeAt(peg$currPos) === 33) {
-        s1 = peg$c18;
+        s1 = peg$c15;
         peg$currPos++;
       } else {
         s1 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c19); }
+        if (peg$silentFails === 0) { peg$fail(peg$c16); }
       }
       if (s1 !== peg$FAILED) {
         s2 = peg$parse__();
         if (s2 !== peg$FAILED) {
           s3 = peg$parseNotExpr();
           if (s3 !== peg$FAILED) {
-            peg$reportedPos = s0;
-            s1 = peg$c20(s3);
+            peg$savedPos = s0;
+            s1 = peg$c17(s3);
             s0 = s1;
           } else {
             peg$currPos = s0;
-            s0 = peg$c1;
+            s0 = peg$FAILED;
           }
         } else {
           peg$currPos = s0;
-          s0 = peg$c1;
+          s0 = peg$FAILED;
         }
       } else {
         peg$currPos = s0;
-        s0 = peg$c1;
+        s0 = peg$FAILED;
       }
       if (s0 === peg$FAILED) {
         s0 = peg$parseBindingExpr();
@@ -585,11 +594,11 @@ module.exports = (function() {
 
       s0 = peg$currPos;
       if (input.charCodeAt(peg$currPos) === 40) {
-        s1 = peg$c21;
+        s1 = peg$c18;
         peg$currPos++;
       } else {
         s1 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c22); }
+        if (peg$silentFails === 0) { peg$fail(peg$c19); }
       }
       if (s1 !== peg$FAILED) {
         s2 = peg$parse__();
@@ -599,35 +608,35 @@ module.exports = (function() {
             s4 = peg$parse__();
             if (s4 !== peg$FAILED) {
               if (input.charCodeAt(peg$currPos) === 41) {
-                s5 = peg$c23;
+                s5 = peg$c20;
                 peg$currPos++;
               } else {
                 s5 = peg$FAILED;
-                if (peg$silentFails === 0) { peg$fail(peg$c24); }
+                if (peg$silentFails === 0) { peg$fail(peg$c21); }
               }
               if (s5 !== peg$FAILED) {
-                peg$reportedPos = s0;
-                s1 = peg$c25(s3);
+                peg$savedPos = s0;
+                s1 = peg$c22(s3);
                 s0 = s1;
               } else {
                 peg$currPos = s0;
-                s0 = peg$c1;
+                s0 = peg$FAILED;
               }
             } else {
               peg$currPos = s0;
-              s0 = peg$c1;
+              s0 = peg$FAILED;
             }
           } else {
             peg$currPos = s0;
-            s0 = peg$c1;
+            s0 = peg$FAILED;
           }
         } else {
           peg$currPos = s0;
-          s0 = peg$c1;
+          s0 = peg$FAILED;
         }
       } else {
         peg$currPos = s0;
-        s0 = peg$c1;
+        s0 = peg$FAILED;
       }
       if (s0 === peg$FAILED) {
         s0 = peg$parseExpr();
@@ -653,58 +662,58 @@ module.exports = (function() {
       s0 = peg$parseBooleanLiteral();
       if (s0 === peg$FAILED) {
         s0 = peg$currPos;
-        if (input.substr(peg$currPos, 2) === peg$c26) {
-          s1 = peg$c26;
+        if (input.substr(peg$currPos, 2) === peg$c23) {
+          s1 = peg$c23;
           peg$currPos += 2;
         } else {
           s1 = peg$FAILED;
-          if (peg$silentFails === 0) { peg$fail(peg$c27); }
+          if (peg$silentFails === 0) { peg$fail(peg$c24); }
         }
         if (s1 !== peg$FAILED) {
-          peg$reportedPos = s0;
-          s1 = peg$c28();
+          peg$savedPos = s0;
+          s1 = peg$c25();
         }
         s0 = s1;
         if (s0 === peg$FAILED) {
           s0 = peg$currPos;
-          if (input.substr(peg$currPos, 2) === peg$c29) {
-            s1 = peg$c29;
+          if (input.substr(peg$currPos, 2) === peg$c26) {
+            s1 = peg$c26;
             peg$currPos += 2;
           } else {
             s1 = peg$FAILED;
-            if (peg$silentFails === 0) { peg$fail(peg$c30); }
+            if (peg$silentFails === 0) { peg$fail(peg$c27); }
           }
           if (s1 !== peg$FAILED) {
-            peg$reportedPos = s0;
-            s1 = peg$c31();
+            peg$savedPos = s0;
+            s1 = peg$c28();
           }
           s0 = s1;
           if (s0 === peg$FAILED) {
             s0 = peg$currPos;
-            if (input.substr(peg$currPos, 2) === peg$c32) {
-              s1 = peg$c32;
+            if (input.substr(peg$currPos, 2) === peg$c29) {
+              s1 = peg$c29;
               peg$currPos += 2;
             } else {
               s1 = peg$FAILED;
-              if (peg$silentFails === 0) { peg$fail(peg$c33); }
+              if (peg$silentFails === 0) { peg$fail(peg$c30); }
             }
             if (s1 !== peg$FAILED) {
-              peg$reportedPos = s0;
-              s1 = peg$c34();
+              peg$savedPos = s0;
+              s1 = peg$c31();
             }
             s0 = s1;
             if (s0 === peg$FAILED) {
               s0 = peg$currPos;
-              if (input.substr(peg$currPos, 2) === peg$c35) {
-                s1 = peg$c35;
+              if (input.substr(peg$currPos, 2) === peg$c32) {
+                s1 = peg$c32;
                 peg$currPos += 2;
               } else {
                 s1 = peg$FAILED;
-                if (peg$silentFails === 0) { peg$fail(peg$c36); }
+                if (peg$silentFails === 0) { peg$fail(peg$c33); }
               }
               if (s1 !== peg$FAILED) {
-                peg$reportedPos = s0;
-                s1 = peg$c37();
+                peg$savedPos = s0;
+                s1 = peg$c34();
               }
               s0 = s1;
             }
@@ -719,30 +728,30 @@ module.exports = (function() {
       var s0, s1;
 
       s0 = peg$currPos;
-      if (input.substr(peg$currPos, 4) === peg$c38) {
-        s1 = peg$c38;
+      if (input.substr(peg$currPos, 4) === peg$c35) {
+        s1 = peg$c35;
         peg$currPos += 4;
       } else {
         s1 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c39); }
+        if (peg$silentFails === 0) { peg$fail(peg$c36); }
       }
       if (s1 !== peg$FAILED) {
-        peg$reportedPos = s0;
-        s1 = peg$c40();
+        peg$savedPos = s0;
+        s1 = peg$c37();
       }
       s0 = s1;
       if (s0 === peg$FAILED) {
         s0 = peg$currPos;
-        if (input.substr(peg$currPos, 5) === peg$c41) {
-          s1 = peg$c41;
+        if (input.substr(peg$currPos, 5) === peg$c38) {
+          s1 = peg$c38;
           peg$currPos += 5;
         } else {
           s1 = peg$FAILED;
-          if (peg$silentFails === 0) { peg$fail(peg$c42); }
+          if (peg$silentFails === 0) { peg$fail(peg$c39); }
         }
         if (s1 !== peg$FAILED) {
-          peg$reportedPos = s0;
-          s1 = peg$c43();
+          peg$savedPos = s0;
+          s1 = peg$c40();
         }
         s0 = s1;
       }
@@ -754,12 +763,12 @@ module.exports = (function() {
       var s0, s1, s2, s3;
 
       s0 = peg$currPos;
-      if (input.substr(peg$currPos, 2) === peg$c44) {
-        s1 = peg$c44;
+      if (input.substr(peg$currPos, 2) === peg$c41) {
+        s1 = peg$c41;
         peg$currPos += 2;
       } else {
         s1 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c45); }
+        if (peg$silentFails === 0) { peg$fail(peg$c42); }
       }
       if (s1 !== peg$FAILED) {
         s2 = [];
@@ -770,34 +779,34 @@ module.exports = (function() {
             s3 = peg$parsews();
           }
         } else {
-          s2 = peg$c1;
+          s2 = peg$FAILED;
         }
         if (s2 !== peg$FAILED) {
           s3 = peg$parseIntegerLiteral();
           if (s3 !== peg$FAILED) {
-            peg$reportedPos = s0;
-            s1 = peg$c46(s3);
+            peg$savedPos = s0;
+            s1 = peg$c43(s3);
             s0 = s1;
           } else {
             peg$currPos = s0;
-            s0 = peg$c1;
+            s0 = peg$FAILED;
           }
         } else {
           peg$currPos = s0;
-          s0 = peg$c1;
+          s0 = peg$FAILED;
         }
       } else {
         peg$currPos = s0;
-        s0 = peg$c1;
+        s0 = peg$FAILED;
       }
       if (s0 === peg$FAILED) {
         s0 = peg$currPos;
-        if (input.substr(peg$currPos, 2) === peg$c47) {
-          s1 = peg$c47;
+        if (input.substr(peg$currPos, 2) === peg$c44) {
+          s1 = peg$c44;
           peg$currPos += 2;
         } else {
           s1 = peg$FAILED;
-          if (peg$silentFails === 0) { peg$fail(peg$c48); }
+          if (peg$silentFails === 0) { peg$fail(peg$c45); }
         }
         if (s1 !== peg$FAILED) {
           s2 = [];
@@ -808,34 +817,34 @@ module.exports = (function() {
               s3 = peg$parsews();
             }
           } else {
-            s2 = peg$c1;
+            s2 = peg$FAILED;
           }
           if (s2 !== peg$FAILED) {
             s3 = peg$parseStringLiteral();
             if (s3 !== peg$FAILED) {
-              peg$reportedPos = s0;
-              s1 = peg$c49(s3);
+              peg$savedPos = s0;
+              s1 = peg$c46(s3);
               s0 = s1;
             } else {
               peg$currPos = s0;
-              s0 = peg$c1;
+              s0 = peg$FAILED;
             }
           } else {
             peg$currPos = s0;
-            s0 = peg$c1;
+            s0 = peg$FAILED;
           }
         } else {
           peg$currPos = s0;
-          s0 = peg$c1;
+          s0 = peg$FAILED;
         }
         if (s0 === peg$FAILED) {
           s0 = peg$currPos;
-          if (input.substr(peg$currPos, 2) === peg$c50) {
-            s1 = peg$c50;
+          if (input.substr(peg$currPos, 2) === peg$c47) {
+            s1 = peg$c47;
             peg$currPos += 2;
           } else {
             s1 = peg$FAILED;
-            if (peg$silentFails === 0) { peg$fail(peg$c51); }
+            if (peg$silentFails === 0) { peg$fail(peg$c48); }
           }
           if (s1 !== peg$FAILED) {
             s2 = [];
@@ -846,34 +855,34 @@ module.exports = (function() {
                 s3 = peg$parsews();
               }
             } else {
-              s2 = peg$c1;
+              s2 = peg$FAILED;
             }
             if (s2 !== peg$FAILED) {
               s3 = peg$parseStringLiteral();
               if (s3 !== peg$FAILED) {
-                peg$reportedPos = s0;
-                s1 = peg$c52(s3);
+                peg$savedPos = s0;
+                s1 = peg$c49(s3);
                 s0 = s1;
               } else {
                 peg$currPos = s0;
-                s0 = peg$c1;
+                s0 = peg$FAILED;
               }
             } else {
               peg$currPos = s0;
-              s0 = peg$c1;
+              s0 = peg$FAILED;
             }
           } else {
             peg$currPos = s0;
-            s0 = peg$c1;
+            s0 = peg$FAILED;
           }
           if (s0 === peg$FAILED) {
             s0 = peg$currPos;
-            if (input.substr(peg$currPos, 3) === peg$c53) {
-              s1 = peg$c53;
+            if (input.substr(peg$currPos, 3) === peg$c50) {
+              s1 = peg$c50;
               peg$currPos += 3;
             } else {
               s1 = peg$FAILED;
-              if (peg$silentFails === 0) { peg$fail(peg$c54); }
+              if (peg$silentFails === 0) { peg$fail(peg$c51); }
             }
             if (s1 !== peg$FAILED) {
               s2 = [];
@@ -884,34 +893,34 @@ module.exports = (function() {
                   s3 = peg$parsews();
                 }
               } else {
-                s2 = peg$c1;
+                s2 = peg$FAILED;
               }
               if (s2 !== peg$FAILED) {
                 s3 = peg$parseStringLiteral();
                 if (s3 !== peg$FAILED) {
-                  peg$reportedPos = s0;
-                  s1 = peg$c55(s3);
+                  peg$savedPos = s0;
+                  s1 = peg$c52(s3);
                   s0 = s1;
                 } else {
                   peg$currPos = s0;
-                  s0 = peg$c1;
+                  s0 = peg$FAILED;
                 }
               } else {
                 peg$currPos = s0;
-                s0 = peg$c1;
+                s0 = peg$FAILED;
               }
             } else {
               peg$currPos = s0;
-              s0 = peg$c1;
+              s0 = peg$FAILED;
             }
             if (s0 === peg$FAILED) {
               s0 = peg$currPos;
-              if (input.substr(peg$currPos, 3) === peg$c56) {
-                s1 = peg$c56;
+              if (input.substr(peg$currPos, 3) === peg$c53) {
+                s1 = peg$c53;
                 peg$currPos += 3;
               } else {
                 s1 = peg$FAILED;
-                if (peg$silentFails === 0) { peg$fail(peg$c57); }
+                if (peg$silentFails === 0) { peg$fail(peg$c54); }
               }
               if (s1 !== peg$FAILED) {
                 s2 = [];
@@ -922,34 +931,34 @@ module.exports = (function() {
                     s3 = peg$parsews();
                   }
                 } else {
-                  s2 = peg$c1;
+                  s2 = peg$FAILED;
                 }
                 if (s2 !== peg$FAILED) {
                   s3 = peg$parseStringLiteral();
                   if (s3 !== peg$FAILED) {
-                    peg$reportedPos = s0;
-                    s1 = peg$c58(s3);
+                    peg$savedPos = s0;
+                    s1 = peg$c55(s3);
                     s0 = s1;
                   } else {
                     peg$currPos = s0;
-                    s0 = peg$c1;
+                    s0 = peg$FAILED;
                   }
                 } else {
                   peg$currPos = s0;
-                  s0 = peg$c1;
+                  s0 = peg$FAILED;
                 }
               } else {
                 peg$currPos = s0;
-                s0 = peg$c1;
+                s0 = peg$FAILED;
               }
               if (s0 === peg$FAILED) {
                 s0 = peg$currPos;
-                if (input.substr(peg$currPos, 2) === peg$c59) {
-                  s1 = peg$c59;
+                if (input.substr(peg$currPos, 2) === peg$c56) {
+                  s1 = peg$c56;
                   peg$currPos += 2;
                 } else {
                   s1 = peg$FAILED;
-                  if (peg$silentFails === 0) { peg$fail(peg$c60); }
+                  if (peg$silentFails === 0) { peg$fail(peg$c57); }
                 }
                 if (s1 !== peg$FAILED) {
                   s2 = [];
@@ -960,34 +969,34 @@ module.exports = (function() {
                       s3 = peg$parsews();
                     }
                   } else {
-                    s2 = peg$c1;
+                    s2 = peg$FAILED;
                   }
                   if (s2 !== peg$FAILED) {
                     s3 = peg$parseStringLiteral();
                     if (s3 !== peg$FAILED) {
-                      peg$reportedPos = s0;
-                      s1 = peg$c61(s3);
+                      peg$savedPos = s0;
+                      s1 = peg$c58(s3);
                       s0 = s1;
                     } else {
                       peg$currPos = s0;
-                      s0 = peg$c1;
+                      s0 = peg$FAILED;
                     }
                   } else {
                     peg$currPos = s0;
-                    s0 = peg$c1;
+                    s0 = peg$FAILED;
                   }
                 } else {
                   peg$currPos = s0;
-                  s0 = peg$c1;
+                  s0 = peg$FAILED;
                 }
                 if (s0 === peg$FAILED) {
                   s0 = peg$currPos;
-                  if (input.substr(peg$currPos, 2) === peg$c62) {
-                    s1 = peg$c62;
+                  if (input.substr(peg$currPos, 2) === peg$c59) {
+                    s1 = peg$c59;
                     peg$currPos += 2;
                   } else {
                     s1 = peg$FAILED;
-                    if (peg$silentFails === 0) { peg$fail(peg$c63); }
+                    if (peg$silentFails === 0) { peg$fail(peg$c60); }
                   }
                   if (s1 !== peg$FAILED) {
                     s2 = [];
@@ -998,34 +1007,34 @@ module.exports = (function() {
                         s3 = peg$parsews();
                       }
                     } else {
-                      s2 = peg$c1;
+                      s2 = peg$FAILED;
                     }
                     if (s2 !== peg$FAILED) {
                       s3 = peg$parseStringLiteral();
                       if (s3 !== peg$FAILED) {
-                        peg$reportedPos = s0;
-                        s1 = peg$c64(s3);
+                        peg$savedPos = s0;
+                        s1 = peg$c61(s3);
                         s0 = s1;
                       } else {
                         peg$currPos = s0;
-                        s0 = peg$c1;
+                        s0 = peg$FAILED;
                       }
                     } else {
                       peg$currPos = s0;
-                      s0 = peg$c1;
+                      s0 = peg$FAILED;
                     }
                   } else {
                     peg$currPos = s0;
-                    s0 = peg$c1;
+                    s0 = peg$FAILED;
                   }
                   if (s0 === peg$FAILED) {
                     s0 = peg$currPos;
-                    if (input.substr(peg$currPos, 3) === peg$c65) {
-                      s1 = peg$c65;
+                    if (input.substr(peg$currPos, 3) === peg$c62) {
+                      s1 = peg$c62;
                       peg$currPos += 3;
                     } else {
                       s1 = peg$FAILED;
-                      if (peg$silentFails === 0) { peg$fail(peg$c66); }
+                      if (peg$silentFails === 0) { peg$fail(peg$c63); }
                     }
                     if (s1 !== peg$FAILED) {
                       s2 = [];
@@ -1036,34 +1045,34 @@ module.exports = (function() {
                           s3 = peg$parsews();
                         }
                       } else {
-                        s2 = peg$c1;
+                        s2 = peg$FAILED;
                       }
                       if (s2 !== peg$FAILED) {
                         s3 = peg$parseStringLiteral();
                         if (s3 !== peg$FAILED) {
-                          peg$reportedPos = s0;
-                          s1 = peg$c67(s3);
+                          peg$savedPos = s0;
+                          s1 = peg$c64(s3);
                           s0 = s1;
                         } else {
                           peg$currPos = s0;
-                          s0 = peg$c1;
+                          s0 = peg$FAILED;
                         }
                       } else {
                         peg$currPos = s0;
-                        s0 = peg$c1;
+                        s0 = peg$FAILED;
                       }
                     } else {
                       peg$currPos = s0;
-                      s0 = peg$c1;
+                      s0 = peg$FAILED;
                     }
                     if (s0 === peg$FAILED) {
                       s0 = peg$currPos;
-                      if (input.substr(peg$currPos, 3) === peg$c68) {
-                        s1 = peg$c68;
+                      if (input.substr(peg$currPos, 3) === peg$c65) {
+                        s1 = peg$c65;
                         peg$currPos += 3;
                       } else {
                         s1 = peg$FAILED;
-                        if (peg$silentFails === 0) { peg$fail(peg$c69); }
+                        if (peg$silentFails === 0) { peg$fail(peg$c66); }
                       }
                       if (s1 !== peg$FAILED) {
                         s2 = [];
@@ -1074,34 +1083,34 @@ module.exports = (function() {
                             s3 = peg$parsews();
                           }
                         } else {
-                          s2 = peg$c1;
+                          s2 = peg$FAILED;
                         }
                         if (s2 !== peg$FAILED) {
                           s3 = peg$parseStringLiteral();
                           if (s3 !== peg$FAILED) {
-                            peg$reportedPos = s0;
-                            s1 = peg$c70(s3);
+                            peg$savedPos = s0;
+                            s1 = peg$c67(s3);
                             s0 = s1;
                           } else {
                             peg$currPos = s0;
-                            s0 = peg$c1;
+                            s0 = peg$FAILED;
                           }
                         } else {
                           peg$currPos = s0;
-                          s0 = peg$c1;
+                          s0 = peg$FAILED;
                         }
                       } else {
                         peg$currPos = s0;
-                        s0 = peg$c1;
+                        s0 = peg$FAILED;
                       }
                       if (s0 === peg$FAILED) {
                         s0 = peg$currPos;
-                        if (input.substr(peg$currPos, 2) === peg$c71) {
-                          s1 = peg$c71;
+                        if (input.substr(peg$currPos, 2) === peg$c68) {
+                          s1 = peg$c68;
                           peg$currPos += 2;
                         } else {
                           s1 = peg$FAILED;
-                          if (peg$silentFails === 0) { peg$fail(peg$c72); }
+                          if (peg$silentFails === 0) { peg$fail(peg$c69); }
                         }
                         if (s1 !== peg$FAILED) {
                           s2 = [];
@@ -1112,32 +1121,32 @@ module.exports = (function() {
                               s3 = peg$parsews();
                             }
                           } else {
-                            s2 = peg$c1;
+                            s2 = peg$FAILED;
                           }
                           if (s2 !== peg$FAILED) {
                             s3 = peg$parseStringLiteral();
                             if (s3 !== peg$FAILED) {
-                              peg$reportedPos = s0;
-                              s1 = peg$c73(s3);
+                              peg$savedPos = s0;
+                              s1 = peg$c70(s3);
                               s0 = s1;
                             } else {
                               peg$currPos = s0;
-                              s0 = peg$c1;
+                              s0 = peg$FAILED;
                             }
                           } else {
                             peg$currPos = s0;
-                            s0 = peg$c1;
+                            s0 = peg$FAILED;
                           }
                         } else {
                           peg$currPos = s0;
-                          s0 = peg$c1;
+                          s0 = peg$FAILED;
                         }
                         if (s0 === peg$FAILED) {
                           s0 = peg$currPos;
                           s1 = peg$parseStringLiteral();
                           if (s1 !== peg$FAILED) {
-                            peg$reportedPos = s0;
-                            s1 = peg$c73(s1);
+                            peg$savedPos = s0;
+                            s1 = peg$c70(s1);
                           }
                           s0 = s1;
                         }
@@ -1159,70 +1168,70 @@ module.exports = (function() {
 
       peg$silentFails++;
       s0 = peg$currPos;
-      if (peg$c76.test(input.charAt(peg$currPos))) {
+      if (peg$c72.test(input.charAt(peg$currPos))) {
         s1 = input.charAt(peg$currPos);
         peg$currPos++;
       } else {
         s1 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c77); }
+        if (peg$silentFails === 0) { peg$fail(peg$c73); }
       }
       if (s1 === peg$FAILED) {
-        s1 = peg$c75;
+        s1 = null;
       }
       if (s1 !== peg$FAILED) {
         s2 = [];
-        if (peg$c78.test(input.charAt(peg$currPos))) {
+        if (peg$c74.test(input.charAt(peg$currPos))) {
           s3 = input.charAt(peg$currPos);
           peg$currPos++;
         } else {
           s3 = peg$FAILED;
-          if (peg$silentFails === 0) { peg$fail(peg$c79); }
+          if (peg$silentFails === 0) { peg$fail(peg$c75); }
         }
         if (s3 !== peg$FAILED) {
           while (s3 !== peg$FAILED) {
             s2.push(s3);
-            if (peg$c78.test(input.charAt(peg$currPos))) {
+            if (peg$c74.test(input.charAt(peg$currPos))) {
               s3 = input.charAt(peg$currPos);
               peg$currPos++;
             } else {
               s3 = peg$FAILED;
-              if (peg$silentFails === 0) { peg$fail(peg$c79); }
+              if (peg$silentFails === 0) { peg$fail(peg$c75); }
             }
           }
         } else {
-          s2 = peg$c1;
+          s2 = peg$FAILED;
         }
         if (s2 !== peg$FAILED) {
-          if (peg$c76.test(input.charAt(peg$currPos))) {
+          if (peg$c72.test(input.charAt(peg$currPos))) {
             s3 = input.charAt(peg$currPos);
             peg$currPos++;
           } else {
             s3 = peg$FAILED;
-            if (peg$silentFails === 0) { peg$fail(peg$c77); }
+            if (peg$silentFails === 0) { peg$fail(peg$c73); }
           }
           if (s3 === peg$FAILED) {
-            s3 = peg$c75;
+            s3 = null;
           }
           if (s3 !== peg$FAILED) {
-            peg$reportedPos = s0;
-            s1 = peg$c80(s2);
+            peg$savedPos = s0;
+            s1 = peg$c76(s2);
             s0 = s1;
           } else {
             peg$currPos = s0;
-            s0 = peg$c1;
+            s0 = peg$FAILED;
           }
         } else {
           peg$currPos = s0;
-          s0 = peg$c1;
+          s0 = peg$FAILED;
         }
       } else {
         peg$currPos = s0;
-        s0 = peg$c1;
+        s0 = peg$FAILED;
       }
       peg$silentFails--;
       if (s0 === peg$FAILED) {
         s1 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c74); }
+        if (peg$silentFails === 0) { peg$fail(peg$c71); }
       }
 
       return s0;
@@ -1234,11 +1243,11 @@ module.exports = (function() {
       peg$silentFails++;
       s0 = peg$currPos;
       if (input.charCodeAt(peg$currPos) === 34) {
-        s1 = peg$c82;
+        s1 = peg$c78;
         peg$currPos++;
       } else {
         s1 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c83); }
+        if (peg$silentFails === 0) { peg$fail(peg$c79); }
       }
       if (s1 !== peg$FAILED) {
         s2 = [];
@@ -1249,36 +1258,36 @@ module.exports = (function() {
         }
         if (s2 !== peg$FAILED) {
           if (input.charCodeAt(peg$currPos) === 34) {
-            s3 = peg$c82;
+            s3 = peg$c78;
             peg$currPos++;
           } else {
             s3 = peg$FAILED;
-            if (peg$silentFails === 0) { peg$fail(peg$c83); }
+            if (peg$silentFails === 0) { peg$fail(peg$c79); }
           }
           if (s3 !== peg$FAILED) {
-            peg$reportedPos = s0;
-            s1 = peg$c84(s2);
+            peg$savedPos = s0;
+            s1 = peg$c80(s2);
             s0 = s1;
           } else {
             peg$currPos = s0;
-            s0 = peg$c1;
+            s0 = peg$FAILED;
           }
         } else {
           peg$currPos = s0;
-          s0 = peg$c1;
+          s0 = peg$FAILED;
         }
       } else {
         peg$currPos = s0;
-        s0 = peg$c1;
+        s0 = peg$FAILED;
       }
       if (s0 === peg$FAILED) {
         s0 = peg$currPos;
         if (input.charCodeAt(peg$currPos) === 39) {
-          s1 = peg$c85;
+          s1 = peg$c81;
           peg$currPos++;
         } else {
           s1 = peg$FAILED;
-          if (peg$silentFails === 0) { peg$fail(peg$c86); }
+          if (peg$silentFails === 0) { peg$fail(peg$c82); }
         }
         if (s1 !== peg$FAILED) {
           s2 = [];
@@ -1289,27 +1298,27 @@ module.exports = (function() {
           }
           if (s2 !== peg$FAILED) {
             if (input.charCodeAt(peg$currPos) === 39) {
-              s3 = peg$c85;
+              s3 = peg$c81;
               peg$currPos++;
             } else {
               s3 = peg$FAILED;
-              if (peg$silentFails === 0) { peg$fail(peg$c86); }
+              if (peg$silentFails === 0) { peg$fail(peg$c82); }
             }
             if (s3 !== peg$FAILED) {
-              peg$reportedPos = s0;
-              s1 = peg$c84(s2);
+              peg$savedPos = s0;
+              s1 = peg$c80(s2);
               s0 = s1;
             } else {
               peg$currPos = s0;
-              s0 = peg$c1;
+              s0 = peg$FAILED;
             }
           } else {
             peg$currPos = s0;
-            s0 = peg$c1;
+            s0 = peg$FAILED;
           }
         } else {
           peg$currPos = s0;
-          s0 = peg$c1;
+          s0 = peg$FAILED;
         }
         if (s0 === peg$FAILED) {
           s0 = peg$currPos;
@@ -1318,10 +1327,10 @@ module.exports = (function() {
           s2 = peg$parsecc();
           peg$silentFails--;
           if (s2 === peg$FAILED) {
-            s1 = peg$c87;
+            s1 = void 0;
           } else {
             peg$currPos = s1;
-            s1 = peg$c1;
+            s1 = peg$FAILED;
           }
           if (s1 !== peg$FAILED) {
             s2 = [];
@@ -1332,26 +1341,26 @@ module.exports = (function() {
                 s3 = peg$parseUnquotedStringChar();
               }
             } else {
-              s2 = peg$c1;
+              s2 = peg$FAILED;
             }
             if (s2 !== peg$FAILED) {
-              peg$reportedPos = s0;
-              s1 = peg$c84(s2);
+              peg$savedPos = s0;
+              s1 = peg$c80(s2);
               s0 = s1;
             } else {
               peg$currPos = s0;
-              s0 = peg$c1;
+              s0 = peg$FAILED;
             }
           } else {
             peg$currPos = s0;
-            s0 = peg$c1;
+            s0 = peg$FAILED;
           }
         }
       }
       peg$silentFails--;
       if (s0 === peg$FAILED) {
         s1 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c81); }
+        if (peg$silentFails === 0) { peg$fail(peg$c77); }
       }
 
       return s0;
@@ -1363,19 +1372,19 @@ module.exports = (function() {
       s0 = peg$currPos;
       s1 = peg$currPos;
       peg$silentFails++;
-      if (peg$c88.test(input.charAt(peg$currPos))) {
+      if (peg$c83.test(input.charAt(peg$currPos))) {
         s2 = input.charAt(peg$currPos);
         peg$currPos++;
       } else {
         s2 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c89); }
+        if (peg$silentFails === 0) { peg$fail(peg$c84); }
       }
       peg$silentFails--;
       if (s2 === peg$FAILED) {
-        s1 = peg$c87;
+        s1 = void 0;
       } else {
         peg$currPos = s1;
-        s1 = peg$c1;
+        s1 = peg$FAILED;
       }
       if (s1 !== peg$FAILED) {
         if (input.length > peg$currPos) {
@@ -1383,42 +1392,42 @@ module.exports = (function() {
           peg$currPos++;
         } else {
           s2 = peg$FAILED;
-          if (peg$silentFails === 0) { peg$fail(peg$c90); }
+          if (peg$silentFails === 0) { peg$fail(peg$c85); }
         }
         if (s2 !== peg$FAILED) {
-          peg$reportedPos = s0;
-          s1 = peg$c91(s2);
+          peg$savedPos = s0;
+          s1 = peg$c86(s2);
           s0 = s1;
         } else {
           peg$currPos = s0;
-          s0 = peg$c1;
+          s0 = peg$FAILED;
         }
       } else {
         peg$currPos = s0;
-        s0 = peg$c1;
+        s0 = peg$FAILED;
       }
       if (s0 === peg$FAILED) {
         s0 = peg$currPos;
         if (input.charCodeAt(peg$currPos) === 92) {
-          s1 = peg$c92;
+          s1 = peg$c87;
           peg$currPos++;
         } else {
           s1 = peg$FAILED;
-          if (peg$silentFails === 0) { peg$fail(peg$c93); }
+          if (peg$silentFails === 0) { peg$fail(peg$c88); }
         }
         if (s1 !== peg$FAILED) {
           s2 = peg$parseEscapeSequence();
           if (s2 !== peg$FAILED) {
-            peg$reportedPos = s0;
-            s1 = peg$c91(s2);
+            peg$savedPos = s0;
+            s1 = peg$c86(s2);
             s0 = s1;
           } else {
             peg$currPos = s0;
-            s0 = peg$c1;
+            s0 = peg$FAILED;
           }
         } else {
           peg$currPos = s0;
-          s0 = peg$c1;
+          s0 = peg$FAILED;
         }
       }
 
@@ -1431,19 +1440,19 @@ module.exports = (function() {
       s0 = peg$currPos;
       s1 = peg$currPos;
       peg$silentFails++;
-      if (peg$c94.test(input.charAt(peg$currPos))) {
+      if (peg$c89.test(input.charAt(peg$currPos))) {
         s2 = input.charAt(peg$currPos);
         peg$currPos++;
       } else {
         s2 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c95); }
+        if (peg$silentFails === 0) { peg$fail(peg$c90); }
       }
       peg$silentFails--;
       if (s2 === peg$FAILED) {
-        s1 = peg$c87;
+        s1 = void 0;
       } else {
         peg$currPos = s1;
-        s1 = peg$c1;
+        s1 = peg$FAILED;
       }
       if (s1 !== peg$FAILED) {
         if (input.length > peg$currPos) {
@@ -1451,42 +1460,42 @@ module.exports = (function() {
           peg$currPos++;
         } else {
           s2 = peg$FAILED;
-          if (peg$silentFails === 0) { peg$fail(peg$c90); }
+          if (peg$silentFails === 0) { peg$fail(peg$c85); }
         }
         if (s2 !== peg$FAILED) {
-          peg$reportedPos = s0;
-          s1 = peg$c91(s2);
+          peg$savedPos = s0;
+          s1 = peg$c86(s2);
           s0 = s1;
         } else {
           peg$currPos = s0;
-          s0 = peg$c1;
+          s0 = peg$FAILED;
         }
       } else {
         peg$currPos = s0;
-        s0 = peg$c1;
+        s0 = peg$FAILED;
       }
       if (s0 === peg$FAILED) {
         s0 = peg$currPos;
         if (input.charCodeAt(peg$currPos) === 92) {
-          s1 = peg$c92;
+          s1 = peg$c87;
           peg$currPos++;
         } else {
           s1 = peg$FAILED;
-          if (peg$silentFails === 0) { peg$fail(peg$c93); }
+          if (peg$silentFails === 0) { peg$fail(peg$c88); }
         }
         if (s1 !== peg$FAILED) {
           s2 = peg$parseEscapeSequence();
           if (s2 !== peg$FAILED) {
-            peg$reportedPos = s0;
-            s1 = peg$c91(s2);
+            peg$savedPos = s0;
+            s1 = peg$c86(s2);
             s0 = s1;
           } else {
             peg$currPos = s0;
-            s0 = peg$c1;
+            s0 = peg$FAILED;
           }
         } else {
           peg$currPos = s0;
-          s0 = peg$c1;
+          s0 = peg$FAILED;
         }
       }
 
@@ -1502,10 +1511,10 @@ module.exports = (function() {
       s2 = peg$parsews();
       peg$silentFails--;
       if (s2 === peg$FAILED) {
-        s1 = peg$c87;
+        s1 = void 0;
       } else {
         peg$currPos = s1;
-        s1 = peg$c1;
+        s1 = peg$FAILED;
       }
       if (s1 !== peg$FAILED) {
         if (input.length > peg$currPos) {
@@ -1513,19 +1522,19 @@ module.exports = (function() {
           peg$currPos++;
         } else {
           s2 = peg$FAILED;
-          if (peg$silentFails === 0) { peg$fail(peg$c90); }
+          if (peg$silentFails === 0) { peg$fail(peg$c85); }
         }
         if (s2 !== peg$FAILED) {
-          peg$reportedPos = s0;
-          s1 = peg$c91(s2);
+          peg$savedPos = s0;
+          s1 = peg$c86(s2);
           s0 = s1;
         } else {
           peg$currPos = s0;
-          s0 = peg$c1;
+          s0 = peg$FAILED;
         }
       } else {
         peg$currPos = s0;
-        s0 = peg$c1;
+        s0 = peg$FAILED;
       }
 
       return s0;
@@ -1534,53 +1543,53 @@ module.exports = (function() {
     function peg$parseEscapeSequence() {
       var s0, s1;
 
-      if (peg$c96.test(input.charAt(peg$currPos))) {
+      if (peg$c91.test(input.charAt(peg$currPos))) {
         s0 = input.charAt(peg$currPos);
         peg$currPos++;
       } else {
         s0 = peg$FAILED;
-        if (peg$silentFails === 0) { peg$fail(peg$c97); }
+        if (peg$silentFails === 0) { peg$fail(peg$c92); }
       }
       if (s0 === peg$FAILED) {
         s0 = peg$currPos;
         if (input.charCodeAt(peg$currPos) === 110) {
-          s1 = peg$c98;
+          s1 = peg$c93;
           peg$currPos++;
         } else {
           s1 = peg$FAILED;
-          if (peg$silentFails === 0) { peg$fail(peg$c99); }
+          if (peg$silentFails === 0) { peg$fail(peg$c94); }
         }
         if (s1 !== peg$FAILED) {
-          peg$reportedPos = s0;
-          s1 = peg$c100();
+          peg$savedPos = s0;
+          s1 = peg$c95();
         }
         s0 = s1;
         if (s0 === peg$FAILED) {
           s0 = peg$currPos;
           if (input.charCodeAt(peg$currPos) === 114) {
-            s1 = peg$c101;
+            s1 = peg$c96;
             peg$currPos++;
           } else {
             s1 = peg$FAILED;
-            if (peg$silentFails === 0) { peg$fail(peg$c102); }
+            if (peg$silentFails === 0) { peg$fail(peg$c97); }
           }
           if (s1 !== peg$FAILED) {
-            peg$reportedPos = s0;
-            s1 = peg$c103();
+            peg$savedPos = s0;
+            s1 = peg$c98();
           }
           s0 = s1;
           if (s0 === peg$FAILED) {
             s0 = peg$currPos;
             if (input.charCodeAt(peg$currPos) === 116) {
-              s1 = peg$c104;
+              s1 = peg$c99;
               peg$currPos++;
             } else {
               s1 = peg$FAILED;
-              if (peg$silentFails === 0) { peg$fail(peg$c105); }
+              if (peg$silentFails === 0) { peg$fail(peg$c100); }
             }
             if (s1 !== peg$FAILED) {
-              peg$reportedPos = s0;
-              s1 = peg$c106();
+              peg$savedPos = s0;
+              s1 = peg$c101();
             }
             s0 = s1;
           }
@@ -1763,12 +1772,19 @@ module.exports = (function() {
         peg$fail({ type: "end", description: "end of input" });
       }
 
-      throw peg$buildException(null, peg$maxFailExpected, peg$maxFailPos);
+      throw peg$buildException(
+        null,
+        peg$maxFailExpected,
+        peg$maxFailPos < input.length ? input.charAt(peg$maxFailPos) : null,
+        peg$maxFailPos < input.length
+          ? peg$computeLocation(peg$maxFailPos, peg$maxFailPos + 1)
+          : peg$computeLocation(peg$maxFailPos, peg$maxFailPos)
+      );
     }
   }
 
   return {
-    SyntaxError: SyntaxError,
-    parse:       parse
+    SyntaxError: peg$SyntaxError,
+    parse:       peg$parse
   };
 })();
